server/python/requirements.txt
# audit_engine — Dependências Python
fastapi>=0.104.0
uvicorn>=0.24.0
polars>=0.20.0
pydantic>=2.5.0
openpyxl>=3.1.0
oracledb>=2.0.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
weasyprint>=63.0
pypdf>=5.1.0
reportlab>=4.2.0
python-docx>=1.1.2
pytest>=8.3.0
